+++ aws/eks-service/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.1.24]() (2024-12-26)
+### Features
+* Add `var.create_kubernetes_namespace` to specify whether to create a namespace
+* Add `var.config_map_env_var_name` to overwrite default kubernetes_config_map name
+* Add `var.secret_env_var_name` to overwrite default kubernetes_secret secret env var name
+
 
 ## [0.1.22]() (2024-12-25)
 ### Features